feat: refactor background service and handlers for Kaspa integration
