+++ api/background/handlers/kaspa/getAccount.ts
