+++ api/background/handlers/kaspa/utils.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+import { ScriptOption } from "@/lib/wallet/wallet-interface.ts";
+
+export const SignTxPayloadSchema = z.object({
+  networkId: z.string(),
+  txJson: z.string(),
+  scripts: z.array(z.custom<ScriptOption>()).optional(),
+});
+
+export type SignTxPayload = z.infer<typeof SignTxPayloadSchema>;
