+++ api/browser.ts
@@ -1,15 +1,10 @@
 import { v4 as uuid } from "uuid";
-import {
-  Action,
-  ApiRequest,
-  ApiResponseSchema,
-  ConnectPayloadSchema,
-  SignTxPayloadSchema,
-  SignMessagePayloadSchema,
-} from "@/api/message";
+import { Action, ApiRequest, ApiResponseSchema } from "@/api/message";
 import { ScriptOption } from "@/lib/wallet/wallet-interface.ts";
 import { EthereumBrowserAPI } from "./ethereum";
-import { ApiUtils } from "@/api/background/utils.ts";
+import { ConnectPayloadSchema } from "@/api/background/handlers/kaspa/connect";
+import { SignTxPayloadSchema } from "@/api/background/handlers/kaspa/utils";
+import { SignMessagePayloadSchema } from "@/api/background/handlers/kaspa/signMessage";
 
 function createApiRequest(
   action: Action,
@@ -28,12 +23,7 @@ function createApiRequest(
 export class KastleBrowserAPI {
   public readonly ethereum = new EthereumBrowserAPI();
 
-  constructor() {
-    window.postMessage(
-      ApiUtils.createApiResponse("kastle_installed", []),
-      window.location.origin,
-    );
-  }
+  constructor() {}
 
   async connect(
     networkId: "mainnet" | "testnet-10" = "mainnet",
