+++ components/screens/browser-api/sign-and-broadcast/LedgerSignAndBroadcast.tsx
@@ -1,4 +1,4 @@
-import { SignTxPayload } from "@/api/message";
+import { SignTxPayload } from "@/api/background/handlers/kaspa/utils";
 import LedgerNotSupported from "@/components/screens/browser-api/sign/LedgerNotSupported";
 import SignAndBroadcast from "@/components/screens/browser-api/sign-and-broadcast/SignAndBroadcast";
 import { AccountFactory } from "@/lib/wallet/wallet-factory";
