+++ components/screens/browser-api/sign-and-broadcast/SignAndBroadcast.tsx
@@ -1,4 +1,4 @@
-import { SignTxPayload } from "@/api/message";
+import { SignTxPayload } from "@/api/background/handlers/kaspa/utils";
 import { ApiExtensionUtils } from "@/api/extension";
 import { IWallet } from "@/lib/wallet/wallet-interface.ts";
 import useWalletManager from "@/hooks/useWalletManager";
