+++ components/screens/browser-api/sign-tx/SignTx.tsx
@@ -1,4 +1,4 @@
-import { ApiResponse, SignTxPayload } from "@/api/message";
+import { SignTxPayload } from "@/api/background/handlers/kaspa/utils";
 import { ApiExtensionUtils } from "@/api/extension";
 import { IWallet } from "@/lib/wallet/wallet-interface.ts";
 import useRpcClientStateful from "@/hooks/useRpcClientStateful";
