+++ components/screens/browser-api/sign/SignConfirm.tsx
@@ -1,4 +1,4 @@
-import { SignTxPayload } from "@/api/message";
+import { SignTxPayload } from "@/api/background/handlers/kaspa/utils";
 import { NetworkType } from "@/contexts/SettingsContext.tsx";
 import useWalletManager from "@/hooks/useWalletManager";
 import {
